# Conflicts:
#	src/shapenodes/SoCube.cpp

+++ src/base/SbImage.cpp
@@ -730,7 +730,7 @@ BOOST_AUTO_TEST_CASE(copyConstruct)
 {
   unsigned char buf [4];
 
-  for (int i=0;i<sizeof(buf); ++i) {
+  for (size_t i=0;i<sizeof(buf); ++i) {
     buf[i]=i;
   }
